refactor(api/platform): drop dead exports, mirror api/app shape
- Delete unused origins.ts (no callers).
- Delete unused internal.ts proof-of-concept and its re-export.
- Trim index.ts to only the exports actually consumed externally
  (PlatformAuthContext/PlatformContext/VerifiedServiceJWT/SERVICE_CALLERS/
  serviceProcedure are package-internal); reorder to match api/app.
- Move co-located tests into src/__tests__/ to mirror api/app layout.
